+++ lib/packages/wikibase/data-model/src/Entity/Item.php
@@ -304,19 +304,19 @@ class Item extends Entity implements StatementListHolder {
 	 *
 	 * @since 0.1
 	 *
-	 * @param EntityDocument $entity
+	 * @param mixed $target
 	 *
 	 * @return bool
 	 */
-	public function equals( EntityDocument $entity ) {
-		if ( $this === $entity ) {
+	public function equals( $target ) {
+		if ( $this === $target ) {
 			return true;
 		}
 
-		return $entity instanceof self
-		       && $this->fingerprint->equals( $entity->fingerprint )
-		       && $this->siteLinks->equals( $entity->siteLinks )
-		       && $this->statements->equals( $entity->statements );
+		return $target instanceof self
+			&& $this->fingerprint->equals( $target->fingerprint )
+			&& $this->siteLinks->equals( $target->siteLinks )
+			&& $this->statements->equals( $target->statements );
 	}
 
 }